requirements.txt
absl-py==0.9.0
attrs==19.3.0
bravado==10.6.0
bravado-core==5.17.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
future==0.18.2
gin-config==0.3.0
gitdb==4.0.5
GitPython==3.1.2
GPUtil==1.4.0
idna==2.9
importlib-metadata==1.6.0
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
kiwisolver==1.2.0
mapextrackt==0.4
matplotlib==3.2.1
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
neptune-client==0.4.113
numpy==1.18.4
oauthlib==3.1.0
packaging==20.3
pandas==1.0.3
Pillow==7.1.2
pkg-resources==0.0.0
psutil==5.7.0
py3nvml==0.2.6
PyJWT==1.7.1
pykeops==1.4
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
simplejson==3.17.0
six==1.14.0
smmap==3.0.4
strict-rfc3339==0.7
swagger-spec-validator==2.5.0
torch==1.4.0+cu100
torchsummary==1.5.1
torchvision==0.5.0+cu100
tqdm==4.46.0
typing-extensions==3.7.4.2
urllib3==1.25.9
webcolors==1.11.1
websocket-client==0.57.0
xmltodict==0.12.0
zipp==3.1.0
